+++ src/routes/ViewHoldingRoute.js
@@ -1,5 +1,10 @@
 import { useContext } from 'react';
-import { useParams } from 'react-router-dom';
+import {
+  useParams,
+  useLocation,
+} from 'react-router-dom';
+
+import { useStripes } from '@folio/stripes/core';
 
 import { DataContext } from '../contexts';
 import ViewHoldingsRecord from '../ViewHoldingsRecord';
@@ -7,10 +12,13 @@ import ViewHoldingsRecord from '../ViewHoldingsRecord';
 const ViewHoldingRoute = () => {
   const { id: instanceId, holdingsrecordid } = useParams();
   const referenceTables = useContext(DataContext);
+  const { okapi } = useStripes();
+  const { state } = useLocation();
 
   return (
     <ViewHoldingsRecord
       id={instanceId}
+      tenantTo={state?.tenantTo || okapi.tenant}
       referenceTables={referenceTables}
       holdingsrecordid={holdingsrecordid}
     />