+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Change history for ui-inventory
 
+## [11.0.0] IN PROGRESS
 
-## [10.1.0] IN PROGRESS
-
+* *BREAKING* Replace imports from quick-marc with stripes-marc-components. Refs UIIN-2636.
 * Make Inventory search and browse query boxes expandable. Refs UIIN-2493.
+* Added support for `containsAny` match option in Advanced search. Refs UIIN-2486.
+* Inventory search/browse: Do not retain checkbox selections when toggling search segment. Refs UIIN-2477.
 
 ## [10.0.1] IN PROGRESS